fixes changes

+++ src/main/java/ru/hogwarts/school/repositories/AvatarRepository.java
@@ -7,6 +7,6 @@ import ru.hogwarts.school.model.Avatar;
 import java.util.Optional;
 
 @Repository
-public interface AvatarRepository extends PagingAndSortingRepository<Avatar, Integer> {
+public interface AvatarRepository extends PagingAndSortingRepository<Avatar, Long> {
     Optional<Avatar> findByStudentId(long id);
 }
